+++ CHANGELOG.md
@@ -12,6 +12,11 @@ on NuGet.org.
 **2018-12-15**
 
 - [x] Initial release
+      
+## 1.1
+**2019-1-21**
+- [x] Support for deserializing from JSON object with `value` property
+- [x] Added supporting packages for using ExtendableEnums with OData